+++ Projects/AVRISP-MKII/Lib/XPROG/TINYNVM.h
@@ -63,8 +63,8 @@
 
 	/* Function Prototypes: */		
 		bool TINYNVM_WaitWhileNVMBusBusy(void);
-		bool TINYNVM_ReadMemory(const uint32_t ReadAddress, uint8_t* ReadBuffer, uint16_t ReadLength);
-		bool TINYNVM_WriteMemory(const uint32_t WriteAddress, const uint8_t* WriteBuffer, uint16_t WriteLength);
+		bool TINYNVM_ReadMemory(const uint16_t ReadAddress, uint8_t* ReadBuffer, uint16_t ReadLength);
+		bool TINYNVM_WriteMemory(const uint16_t WriteAddress, const uint8_t* WriteBuffer, uint16_t WriteLength);
 		bool TINYNVM_EraseMemory(void);
 
 		#if defined(INCLUDE_FROM_TINYNVM_C)
